pyproject.toml
[project]
name = "inatinqperf"
version = "0.0.1" 
#          ^^^^^
# Wheel names will be generated according to this value. Do not manually modify this value; instead
# update it according to committed changes by running this command from the root of the repository:
#
#   uv run python -m AutoGitSemVer.scripts.UpdatePythonVersion ./pyproject.toml ./src

description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ketan Bhardwaj", email = "ketanbj@cc.gatech.edu" }
]
requires-python = ">= 3.11"
dependencies = []
classifiers = [
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
]

[project.license]
text = "MIT"

[project.urls]
Homepage = "https://github.com/gt-sse-center/iNatInqPerf"
Documentation = "https://github.com/gt-sse-center/iNatInqPerf"
Repository = "https://github.com/gt-sse-center/iNatInqPerf"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "autogitsemver>=0.9.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.0",
    "numpy>=1.23,<2.0",
    "pandas>=1.5,<2.3",
    "tqdm>=4.66",
    "pillow>=9.5",
    "psutil>=5.9",
    "datasets>=2.14,<3.0",
    "huggingface-hub>=0.23",
    "safetensors>=0.4",
    "transformers>=4.35,<4.42",
    "torch>=2.0,<2.3",
    "torchvision>=0.15,<0.19",
    "faiss-cpu>=1.7.4",
    "pyyaml>=6.0",
    "codetiming>=1.4",
    "requests>=2.31",
    "py-spy"
]

[tool.pytest.ini_options]
addopts = "--verbose -vv --capture=no --cov=inatinqperf --cov-report html --cov-report term --cov-report xml:coverage.xml --cov-fail-under=95.0"

[tool.scripts]
download = "python src/bench/bench.py download"
embed = "python src/bench/bench.py embed"
build = "python src/bench/bench.py build"
search = "python src/bench/bench.py search"
update = "python src/bench/bench.py update"
run-all = "python src/bench/bench.py run-all --size small --backend faiss.flat"

[tool.ruff]
line-length = 110

[tool.ruff.lint]
exclude = ["tests/**"]

select = ["ALL"]

ignore = [
    "ANN002", # Missing type annotation for `*args`
    "ANN003", # Missing type annotation for `**kwargs`
    "BLE001", # Do not catch blind exception: `Exception`
    "COM812", # Trailing comma missing
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in `__init__` method
    "D202", # No blank lines allowed after function docstring
    "E501", # Line too long
    "FIX002", # Line contains TODO, consider resolving the issue
    "I001", # Import block is un-sorted or un-formatted
    "N802", # Function name `xxx` should be lowercase
    "N999", # Invalid module name
    "RSE102", # Unnecessary parentheses on raise exception
    "S101", # Use of assert detected
    "TC006", # Add quotes to type expression in `typing.cast()`
    "TD002", # Missing author in TODO
    "TD003", # Missing issue link for this TODO
    "TRY002", # Create your own exception
    "TRY300", # Consider moving this statement to an `else` block
    "UP032", # Use f-string instead of `format` call
    "T201", # Use of print
    "INP001", # empty __init__.py file
    "ARG002", # some args needed because of adaptor base class
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.pylint]
max-args = 10
max-branches = 20
max-returns = 20

[tool.uv.build-backend]
module-name = "inatinqperf"
